test(ci): split admin token owner coverage
